+++ bambi/tests/test_built_models.py
@@ -1,3 +1,5 @@
+import logging
+
 import pytest
 
 import arviz as az
@@ -46,6 +48,18 @@ def dm():
 
     data_dir = join(dirname(__file__), "data")
     data = pd.read_csv(join(data_dir, "dm.csv"))
+    return data
+
+
+@pytest.fixture(scope="module")
+def init_data():
+    """
+    Data used to test initialization method
+    """
+    from os.path import dirname, join
+
+    data_dir = join(dirname(__file__), "data")
+    data = pd.read_csv(join(data_dir, "obs.csv"))
     return data
 
 
@@ -552,3 +566,12 @@ def test_potentials():
         "Elemwise{switch,no_inplace}(Elemwise{gt,no_inplace}.0, "
         "TensorConstant{0}, TensorConstant{-inf})"
     )
+
+
+def test_init_fallback(init_data, caplog):
+    model = Model("od ~ temp + (1|source) + 0", init_data)
+    with caplog.at_level(logging.INFO):
+        results = model.fit(draws=100, init="auto")
+        assert "Initializing NUTS using jitter+adapt_diag..." in caplog.text
+        assert "The default initialization" in caplog.text
+        assert "Initializing NUTS using adapt_diag..." in caplog.text